промежуточный коммит

+++ core/orders/ordersController.js
@@ -12,16 +12,21 @@ const keyboardOrders = {
 
 function getOrdersList(bot, chatId) {
     let ordersTemplate = '';
-    database.forEach((order, i) => {
-        // TODO добавить кол-во, если >1  ${order.count}шт.
-        // TODO кнопка УДАЛИТЬ
-        ordersTemplate += `${i+1}. ${order.name}, цена: <b>${order.price}</b>, <a href="/delete ${order.order_id}">УДАЛИТЬ</a> \n`
-    })
-
-    bot.sendMessage(chatId, ordersTemplate, {
-        reply_markup: JSON.stringify(keyboardOrders),
+    let messageParams = {
         parse_mode: 'HTML'
-    })
+    }
+
+    if(database.length > 0) {
+        database.forEach((order, i) => {
+            // TODO кнопка УДАЛИТЬ
+            ordersTemplate += `${i+1}. ${order.name}, цена: <b>${order.price}</b>, <a href="/delete ${order.order_id}">УДАЛИТЬ</a> \n`
+        })
+        messageParams.reply_markup = JSON.stringify(keyboardOrders);
+    } else {
+        ordersTemplate = 'Заказов пока нет';
+    }
+
+    bot.sendMessage(chatId, ordersTemplate, messageParams);
 }
 
 module.exports = {
